first upload of the pure css skeleton with menu's left, top and right
Two versions of the webapp living in the repo: the first and 'naked' version is this.
A new project can be started with the index.html and the webapp_skeleton.css. For sure there some default colors to don't ship a nasty version. The menues are colored with high contrasts.